fix after update Laravel Version
