+++ src/.env.example
@@ -1,6 +1,6 @@
-APP_NAME=Laravel
+APP_NAME=book-game
 APP_ENV=local
-APP_KEY=
+APP_KEY=base64:vf41wZXI05oUlywGtne9MkJZC54w/1yvt74MRWS4bd4=
 APP_DEBUG=true
 APP_URL=http://localhost
 
@@ -8,12 +8,12 @@ LOG_CHANNEL=stack
 LOG_DEPRECATIONS_CHANNEL=null
 LOG_LEVEL=debug
 
-DB_CONNECTION=mysql
-DB_HOST=127.0.0.1
-DB_PORT=3306
-DB_DATABASE=laravel
-DB_USERNAME=root
-DB_PASSWORD=
+DB_CONNECTION=pgsql
+DB_HOST=postgres
+DB_PORT=5432
+DB_DATABASE=book-game
+DB_USERNAME=admin
+DB_PASSWORD=secret
 
 BROADCAST_DRIVER=log
 CACHE_DRIVER=file
@@ -51,7 +51,6 @@ PUSHER_PORT=443
 PUSHER_SCHEME=https
 PUSHER_APP_CLUSTER=mt1
 
-VITE_APP_NAME="${APP_NAME}"
 VITE_PUSHER_APP_KEY="${PUSHER_APP_KEY}"
 VITE_PUSHER_HOST="${PUSHER_HOST}"
 VITE_PUSHER_PORT="${PUSHER_PORT}"
