+++ src/app/Http/Controllers/Admin/EnemyController.php
@@ -5,8 +5,8 @@ namespace App\Http\Controllers\Admin;
 use App\Enums\EnemyAttackTypeEnum;
 use App\Enums\TableNameEnum;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\StoreEnemyRequest;
-use App\Http\Requests\UpdateEnemyRequest;
+use App\Http\Requests\Admin\StoreEnemyRequest;
+use App\Http\Requests\Admin\UpdateEnemyRequest;
 use App\Models\Enemy;
 use DateTime;
 use Exception;
