+++ src/app/Http/Controllers/Controller.php
@@ -10,4 +10,3 @@ class Controller extends BaseController
 {
     use AuthorizesRequests, ValidatesRequests;
 }
-``
