+++ src/app/Http/Requests/Admin/UpdateParagraphRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Admin;
 
 use App\Enums\TableNameEnum;
 use Illuminate\Foundation\Http\FormRequest;
